+++ src/Client/Parser.hs
@@ -11,8 +11,10 @@ module Client.Parser
   , DeployContractOptions (..)
   , clientArgParser
   , parseAddressFromOutput
+  , parseBurncapErrorFromOutput
   , parseContractAddressFromOutput
   , parseSignatureFromOutput
+  , parseSimulationResultFromOutput
   , tzbtcClientAddressParser
   ) where
 
@@ -20,9 +22,10 @@ import Data.Char (isAlpha, isDigit)
 import Fmt (Buildable, pretty)
 import Options.Applicative (eitherReader, help, long, metavar, option, optional, short, str, switch)
 import qualified Options.Applicative as Opt
-import qualified Text.Megaparsec as P (Parsec, customFailure, many, parse, satisfy, skipManyTill)
+import qualified Text.Megaparsec as P
+  (Parsec, customFailure, many, parse, satisfy, skipManyTill)
 import Text.Megaparsec.Char (eol, newline, space, printChar)
-import Text.Megaparsec.Char.Lexer (symbol)
+import Text.Megaparsec.Char.Lexer (float, lexeme, symbol)
 import Text.Megaparsec.Error (ParseErrorBundle, ShowErrorComponent(..))
 import Tezos.Common.Json (TezosInt64)
 
@@ -56,8 +59,6 @@ clientArgParser =
     dryRunSwitch =
       switch (long "dry-run" <>
               help "Dry run command to ensure correctness of the arguments")
-    toMuTez :: Double -> TezosInt64
-    toMuTez x = fromInteger (round (x * 10e6))
 
 clientArgRawParser :: Opt.Parser ClientArgsRaw
 clientArgRawParser = Opt.hsubparser $
@@ -435,3 +436,30 @@ tzbtcClientAddressParser = do
 parseContractAddressFromOutput
   :: Text -> Either (ParseErrorBundle Text OutputParseError) Address
 parseContractAddressFromOutput output = P.parse tzbtcClientAddressParser "" output
+
+parseSimulationResultFromOutput
+  :: Text -> Either (ParseErrorBundle Text OutputParseError) SimulationResult
+parseSimulationResultFromOutput output =
+  P.parse (SimulationResult <$> simulationResultParser) "" output
+
+parseBurncapErrorFromOutput
+  :: Text -> Either (ParseErrorBundle Text OutputParseError) TezosInt64
+parseBurncapErrorFromOutput output =
+  P.parse burnCapParser "" output
+
+burnCapParser :: Parser TezosInt64
+burnCapParser = toMuTez <$> do
+  P.skipManyTill (printChar <|> newline) $ do
+    void $ symbol space "The operation will burn"
+    P.skipManyTill printChar $ lexeme (space >> pure ()) float
+
+simulationResultParser :: Parser TezosInt64
+simulationResultParser = toMuTez <$> do
+  P.skipManyTill (printChar <|> newline) $ do
+    void $ symbol space "Fee to the baker: "
+    P.skipManyTill printChar $ lexeme (newline >> pure ()) float
+
+toMuTez :: Double -> TezosInt64
+toMuTez mt_ = fromInteger $ floor $ mt_ * 10e6
+-- floor does not loose precision here since the fees
+-- from simulation won't have more precision then 1 microtez